requirements.txt
ultralytics==8.0.196
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.24.3
flask==2.3.2
collections-extended==2.0.2